Basic Components
Added basic components to be used in the overall project: Icon (which will use the iconLibrary from the index to render an svg of an icon) and Button (which will create a button based on the type of button specified and can use the Icon component within it to also render the button with the icon).  Current decent documentation for both components... I will consider improving it later.