Fixed animation issue
Fixed an issue where the animation of elements in the svelte each-blocks would not trigger on the item creation because of the block getting loaded in as a whole.  This got fixed with a 'global' tag on the animation.  While troubleshooting this I made some of the variable names easier to deal with, though less readable (i.e., 'cardNum' -> 'index' or 'questionNum' -> 'index')